backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Cors>=4.0
python-dotenv>=1.0
PyJWT>=2.8
pdfplumber>=0.7.6
PyPDF2>=3.0.0
nltk>=3.7
Werkzeug>=2.2
flask
flask-cors
flask-sqlalchemy
python-dotenv
PyJWT
werkzeug
pdfplumber
PyPDF2
nltk
